requirements.txt
black==22.6.0
django~=3.2.15
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
django-allauth
djoser==2.1.0
drf-yasg==1.21.3
isort==5.10.1
python-dotenv